pyproject.toml
[tool.poetry]
name = "support-api"
version = "0.1.0"
description = ""
authors = ["nerdthejohn <newtonjohn043@gmail.com>"]
readme = "README.md"
packages = [{include = "support_api"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.7"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
